requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-Cors==4.0.0
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
python-dotenv==1.0.1
pywin32==306
requests==2.31.0
urllib3==2.2.1
waitress==3.0.0
Werkzeug==3.0.1
